Crated inner static class
